+++ portal/routes_auth.py
@@ -19,7 +19,7 @@ def load_app_client():
 
 @blueprint.route('/login')
 def login():
-    """
+    '''
     Login via Globus Auth.
     May be invoked in one of two scenarios:
 
@@ -27,7 +27,7 @@ def login():
       2. Returning to application during login, already have short-lived
          code from Globus Auth to exchange for tokens, encoded in a query
          param
-    """
+    '''
     # The redirect URI, as a complete URI (not relative path)
     redirect_uri = url_for('routes_auth.login', _external=True)
 
@@ -37,16 +37,27 @@ def login():
     # If there's no "code" query string parameter, we're in this route
     # starting a Globus Auth login flow; Redirect out to Globus Auth:
     if 'code' not in request.args:
-        auth_uri = client.oauth2_get_authorize_url()
+        auth_uri = client.oauth2_get_authorize_url(
+            additional_params={
+                'scope': ' '.join([
+                    'openid profile email',
+                    'urn:globus:auth:scope:transfer.api.globus.org:all',
+                    'urn:globus:auth:scope:auth.globus.org:view_identities',
+                    'urn:globus:auth:scope:nexus.api.globus.org:groups'
+                ])
+            }
+        )
         return redirect(auth_uri)
 
     # If we do have a "code" param, we're coming back from Globus Auth
     # and can start the process of exchanging an auth code for a token.
     code = request.args.get('code')
     tokens = client.oauth2_exchange_code_for_tokens(code)
+    # The repr is deceptive: Looks like a dict, but direct access not possible.
+    nexus_token = tokens.by_resource_server['nexus.api.globus.org']['access_token']
 
     session.update(
-        tokens=tokens.by_resource_server,
+        nexus_token=nexus_token,
         is_authenticated=True
     )
     return redirect(url_for('routes.index'))
@@ -54,11 +65,11 @@ def login():
 
 @blueprint.route('/logout')
 def logout():
-    """
+    '''
     - Revoke the tokens with Globus Auth.
     - Destroy the session state.
     - Redirect the user to the Globus Auth logout page.
-    """
+    '''
     client = load_app_client()
 
     # Revoke the tokens with Globus Auth